Refactor plugin configuration
